+++ src/app/dim-ui/Sheet.tsx
@@ -5,10 +5,12 @@ import { Spring, config, animated } from 'react-spring';
 import { withGesture, GestureState } from 'react-with-gesture';
 import classNames from 'classnames';
 import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
+import * as _ from 'lodash';
 
 interface Props {
-  header?: React.ReactNode;
-  children?: React.ReactNode;
+  header?: React.ReactNode | ((args: { onClose(): void }) => React.ReactNode);
+  children?: React.ReactNode | ((args: { onClose(): void }) => React.ReactNode);
+  sheetClassName?: string;
   onClose(): void;
 }
 
@@ -73,7 +75,7 @@ class Sheet extends React.Component<Props & Partial<GestureState>> {
   }
 
   render() {
-    const { header, children } = this.props;
+    const { header, children, sheetClassName } = this.props;
     const { dragging, closing } = this.state;
 
     const yDelta = closing ? this.height() : dragging ? Math.max(0, this.props.yDelta || 0) : 0;
@@ -93,7 +95,7 @@ class Sheet extends React.Component<Props & Partial<GestureState>> {
         {(style) => (
           <animated.div
             style={{ ...style, maxHeight }}
-            className={classNames('sheet')}
+            className={classNames('sheet', sheetClassName)}
             ref={this.sheet}
             onMouseDown={this.dragHandleDown}
             onMouseUp={this.dragHandleUp}
@@ -109,10 +111,14 @@ class Sheet extends React.Component<Props & Partial<GestureState>> {
             </div>
 
             <div className="sheet-container" style={{ maxHeight }}>
-              {header && <div className="sheet-header">{header}</div>}
+              {header && (
+                <div className="sheet-header">
+                  {_.isFunction(header) ? header({ onClose: this.onClose }) : header}
+                </div>
+              )}
 
               <div className="sheet-contents" ref={this.sheetContents}>
-                {children}
+                {_.isFunction(children) ? children({ onClose: this.onClose }) : children}
               </div>
             </div>
           </animated.div>