+++ src/app/shell/rating-mode/RatingMode.tsx
@@ -11,11 +11,13 @@ import store from '../../store/store';
 import { connect } from 'react-redux';
 import { RootState } from '../../store/reducers';
 import { refresh } from '../refresh';
-import { AppIcon, thumbsUpIcon, uploadIcon } from '../icons';
+import { AppIcon, thumbsUpIcon } from '../icons';
 import { dimCuratedRollService } from '../../curated-rolls/curatedRollService';
 import { updateCurations } from '../../curated-rolls/actions';
 import HelpLink from '../../dim-ui/HelpLink';
 import RatingsKey from '../../item-review/RatingsKey';
+import { DropFilesEventHandler } from 'react-dropzone';
+import FileUpload from '../../dim-ui/FileUpload';
 
 interface StoreProps {
   reviewsModeSelection: number;
@@ -42,8 +44,6 @@ function mapStateToProps(state: RootState): StoreProps {
 class RatingMode extends React.Component<Props, State> {
   private dropdownToggler = React.createRef<HTMLElement>();
   private _reviewModeOptions?: D2ReviewMode[];
-
-  private fileInput = React.createRef<HTMLInputElement>();
 
   constructor(props) {
     super(props);
@@ -81,49 +81,41 @@ class RatingMode extends React.Component<Props, State> {
                 <>
                   <RatingsKey />
                   <div className="mode-row">
-                    <div className="mode-column">
-                      <label className="mode-label" htmlFor="reviewMode">
-                        {t('DtrReview.ForGameMode')}
-                      </label>
-                    </div>
+                    <label className="mode-label" htmlFor="reviewMode">
+                      {t('DtrReview.ForGameMode')}
+                    </label>
                   </div>
                   <div className="mode-row">
-                    <div className="mode-column">
-                      <select
-                        name="reviewMode"
-                        value={reviewsModeSelection}
-                        onChange={this.modeChange}
-                      >
-                        {this.reviewModeOptions.map((r) => (
-                          <option key={r.mode} value={r.mode}>
-                            {r.description}
-                          </option>
-                        ))}
-                      </select>
-                    </div>
+                    <select
+                      name="reviewMode"
+                      value={reviewsModeSelection}
+                      onChange={this.modeChange}
+                    >
+                      {this.reviewModeOptions.map((r) => (
+                        <option key={r.mode} value={r.mode}>
+                          {r.description}
+                        </option>
+                      ))}
+                    </select>
                   </div>
 
                   <div className="mode-row">
-                    <div className="mode-column">
-                      <label className="mode-label" htmlFor="reviewMode">
-                        {t('DtrReview.ForPlatform')}
-                      </label>
-                    </div>
+                    <label className="mode-label" htmlFor="reviewMode">
+                      {t('DtrReview.ForPlatform')}
+                    </label>
                   </div>
                   <div className="mode-row">
-                    <div className="mode-column">
-                      <select
-                        name="platformSelection"
-                        value={platformSelection}
-                        onChange={this.platformChange}
-                      >
-                        {reviewPlatformOptions.map((r) => (
-                          <option key={r.description} value={r.platform}>
-                            {t(r.description)}
-                          </option>
-                        ))}
-                      </select>
-                    </div>
+                    <select
+                      name="platformSelection"
+                      value={platformSelection}
+                      onChange={this.platformChange}
+                    >
+                      {reviewPlatformOptions.map((r) => (
+                        <option key={r.description} value={r.platform}>
+                          {t(r.description)}
+                        </option>
+                      ))}
+                    </select>
                   </div>
                 </>
               )}
@@ -131,21 +123,13 @@ class RatingMode extends React.Component<Props, State> {
               {$featureFlags.curatedRolls && (
                 <>
                   <div className="mode-row">
-                    <div className="mode-column">
-                      <label className="mode-label" htmlFor="curatedRoll">
-                        {t('CuratedRoll.Header')}
-                        <HelpLink helpLink="https://github.com/DestinyItemManager/DIM/blob/master/docs/COMMUNITY_CURATIONS.md" />
-                      </label>
-                    </div>
+                    <label className="mode-label" htmlFor="curatedRoll">
+                      {t('CuratedRoll.Header')}
+                      <HelpLink helpLink="https://github.com/DestinyItemManager/DIM/blob/master/docs/COMMUNITY_CURATIONS.md" />
+                    </label>
                   </div>
                   <div className="mode-row">
-                    <div className="mode-column">
-                      <button className="dim-button" onClick={this.loadCurations}>
-                        <AppIcon icon={uploadIcon} /> <span>{t('CuratedRoll.Import')}</span>
-                      </button>
-                      <br />
-                      <input type="file" id="importFile" ref={this.fileInput} />
-                    </div>
+                    <FileUpload onDrop={this.loadCurations} title={t('CuratedRoll.Import')} />
                   </div>
                 </>
               )}
@@ -195,29 +179,36 @@ class RatingMode extends React.Component<Props, State> {
     refresh();
   };
 
-  private loadCurations = (e) => {
-    e.preventDefault();
+  private loadCurations: DropFilesEventHandler = (acceptedFiles) => {
     const reader = new FileReader();
     reader.onload = () => {
       // TODO: we're kinda trusting that this is the right data here, no validation!
       if (reader.result && typeof reader.result === 'string') {
         dimCuratedRollService.loadCuratedRolls(reader.result);
 
-        const storeRolls = D2StoresService.getStores();
-        const inventoryCuratedRolls = dimCuratedRollService.getInventoryCuratedRolls(storeRolls);
+        if (dimCuratedRollService.getCuratedRolls()) {
+          const storeRolls = D2StoresService.getStores();
+          const inventoryCuratedRolls = dimCuratedRollService.getInventoryCuratedRolls(storeRolls);
 
-        const curationActionData = {
-          curationEnabled: dimCuratedRollService.curationEnabled,
-          inventoryCuratedRolls
-        };
+          const curationActionData = {
+            curationEnabled: dimCuratedRollService.curationEnabled,
+            inventoryCuratedRolls
+          };
 
-        store.dispatch(updateCurations(curationActionData));
-        refresh();
-        alert(t('CuratedRoll.ImportSuccess'));
+          store.dispatch(updateCurations(curationActionData));
+          refresh();
+          alert(
+            t('CuratedRoll.ImportSuccess', {
+              count: dimCuratedRollService.getCuratedRolls().length
+            })
+          );
+        } else {
+          alert(t('CuratedRoll.ImportFailed'));
+        }
       }
     };
 
-    const file = this.fileInput.current!.files![0];
+    const file = acceptedFiles[0];
     if (file) {
       reader.readAsText(file);
     } else {